Move theme related code into theme.py file
